feat: Add framer-motion for animations and implement ContactForm component
